feat: Double-clickable mode for slider
This is to support use cases such as adding additional dots to a slider
on a double click. I made this mutually exclusive with clickable so
that the behavior of clickable can still be to move a dot to a new
position, while the double-clickable option just emits a double-click
event that can then be handled.

This behavior does seem mutually exclusive with single clickable; it
might be better here to have "double" be a value of clickable, since
if a slider is both clickable and double-clickable, the single-click
would move a dot over to the spot being double-clicked.